Add rock-dbg_defconfig which turns on KFD
This file only turns on a substantial smaller set of kernel options,
together with all AMD stuff, and supports most use cases on common
HW configurations. As a result, building kernel with this config
is super fast.

Moreover, with this config, KFD folks can easily build the kernel with
build_kernel.sh script and verify the change before we push the commit
out for review on amd-gfx mailist.

Signed-off-by: Yong Zhao <Yong.Zhao@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>