drm/amd/display: always use 100us for cr aux rd interval
[why]
It is a regression caused by below commit.
I1654c11ba3e229e17483959e0f66319d7de9e320
In this commit the cr training aux rd interval is
modified without following specs requirements.
According to the commit message the change was not intended to modify the value.
Therefore it looks like it is caused by a typo in the change.

Signed-off-by: Wenjing Liu <wenjing.liu@amd.com>
Acked-by: Aurabindo Pillai <aurabindo.pillai@amd.com>